new migration and some feature from Thu

+++ sources/dekidz/database/migrations/2016_09_19_100119_create_table_export_detail.php
@@ -16,7 +16,7 @@ class CreateTableExportDetail extends Migration
             $table->increments('id');
             $table->integer('export_id');
             $table->integer('good_id');
-            $table->integer('quantity');
+            $table->double('quantity', 10, 2);
             $table->timestamps();
         });
     }
